move step & ramp to math_utils
