+++ tests/test_math_utils.py
@@ -1,4 +1,6 @@
 import pytest
+import numpy as np
+import pandas as pd
 import math_utils as mu
 
 
@@ -16,5 +18,41 @@ import math_utils as mu
     (1e12, 1, False),
     (-.001, 0, False),
 ])
-def test_calc_commission(a, b, result):
+def test_isclose(a, b, result):
     assert mu.isclose(a, b) == result
+
+
+@pytest.mark.smoketest
+def test_step():
+    t = np.arange(-5, 5, 1)
+    x = mu.step(t)
+    assert len(x) == len(t)
+    assert x.min() == 0
+    assert x.max() == 1
+
+
+@pytest.mark.smoketest
+def test_step_series():
+    t = pd.Series(np.arange(-5, 5, 1))
+    x = mu.step(t)
+    assert len(x) == len(t)
+    assert x.min() == 0
+    assert x.max() == 1
+
+
+@pytest.mark.smoketest
+def test_ramp():
+    t = np.arange(-5, 5, 1)
+    x = mu.ramp(t)
+    assert len(x) == len(t)
+    assert x.min() == 0
+    assert x.max() == t.max()
+
+
+@pytest.mark.smoketest
+def test_ramp_series():
+    t = pd.Series(np.arange(-5, 5, 1))
+    x = mu.ramp(t)
+    assert len(x) == len(t)
+    assert x.min() == 0
+    assert x.max() == t.max()
